refactor(commands): extract session/output tool commands into session_tools.rs
